refactor: :art: Changed folder structure of tests in the home broker page
